Added some basic tests, moved code from Python to VIMScript and implemented asynchronous search
Clearly Python is not needed at this stage and VIMScript will be needed as we perform the asynchronous job handling stuff it looks like. Can add stuff to Python later for filtering as an example of how to do this.

These seemed to be causing some problems in the Docker container. Removed as it's not that important for now.